[UPNP] Allow disabling UPNP implementation on the Web
Make UPNP classes custom instance, so they can come with empty
implementation on the Web (where the would not work anyway) without
breaking scripts referencing it.

This results in smaller Web builds by not including the library which
also results in the emscripten socket wrappers to be stipped away.
