+++ modules/upnp/register_types.cpp
@@ -33,6 +33,11 @@
 #include "upnp.h"
 #include "upnp_device.h"
 
+#ifndef WEB_ENABLED
+#include "upnp_device_miniupnp.h"
+#include "upnp_miniupnp.h"
+#endif
+
 #include "core/error/error_macros.h"
 
 void initialize_upnp_module(ModuleInitializationLevel p_level) {
@@ -40,8 +45,13 @@ void initialize_upnp_module(ModuleInitializationLevel p_level) {
 		return;
 	}
 
-	GDREGISTER_CLASS(UPNP);
-	GDREGISTER_CLASS(UPNPDevice);
+	ClassDB::register_custom_instance_class<UPNP>();
+	ClassDB::register_custom_instance_class<UPNPDevice>();
+
+#ifndef WEB_ENABLED
+	UPNPMiniUPNP::make_default();
+	UPNPDeviceMiniUPNP::make_default();
+#endif
 }
 
 void uninitialize_upnp_module(ModuleInitializationLevel p_level) {
