+++ modules/upnp/upnp_device_miniupnp.h
@@ -1,5 +1,5 @@
 /**************************************************************************/
-/*  upnp_device.h                                                         */
+/*  upnp_device_miniupnp.h                                                */
 /**************************************************************************/
 /*                         This file is part of:                          */
 /*                             GODOT ENGINE                               */
@@ -28,67 +28,56 @@
 /* SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                 */
 /**************************************************************************/
 
-#ifndef UPNP_DEVICE_H
-#define UPNP_DEVICE_H
+#ifndef UPNP_DEVICE_MINIUPNP_H
+#define UPNP_DEVICE_MINIUPNP_H
 
-#include "core/object/ref_counted.h"
+#ifndef WEB_ENABLED
 
-class UPNPDevice : public RefCounted {
-	GDCLASS(UPNPDevice, RefCounted);
+#include "upnp_device.h"
 
-protected:
-	static void _bind_methods();
+class UPNPDeviceMiniUPNP : public UPNPDevice {
+	GDCLASS(UPNPDeviceMiniUPNP, UPNPDevice);
 
-	static UPNPDevice *(*_create)(bool p_notify_postinitialize);
+private:
+	static UPNPDevice *_create(bool p_notify_postinitialize) { return static_cast<UPNPDevice *>(ClassDB::creator<UPNPDeviceMiniUPNP>(p_notify_postinitialize)); }
+
+	String description_url;
+	String service_type;
+	String igd_control_url;
+	String igd_service_type;
+	String igd_our_addr;
+	IGDStatus igd_status = IGD_STATUS_UNKNOWN_ERROR;
 
 public:
-	enum IGDStatus {
-		IGD_STATUS_OK,
-		IGD_STATUS_HTTP_ERROR,
-		IGD_STATUS_HTTP_EMPTY,
-		IGD_STATUS_NO_URLS,
-		IGD_STATUS_NO_IGD,
-		IGD_STATUS_DISCONNECTED,
-		IGD_STATUS_UNKNOWN_DEVICE,
-		IGD_STATUS_INVALID_CONTROL,
-		IGD_STATUS_MALLOC_ERROR,
-		IGD_STATUS_UNKNOWN_ERROR,
-	};
+	static void make_default();
 
-	static UPNPDevice *create(bool p_notify_postinitialize = true) {
-		if (!_create) {
-			return nullptr;
-		}
-		return _create(p_notify_postinitialize);
-	}
+	virtual void set_description_url(const String &url) override;
+	virtual String get_description_url() const override;
 
-	virtual void set_description_url(const String &url) = 0;
-	virtual String get_description_url() const = 0;
+	virtual void set_service_type(const String &type) override;
+	virtual String get_service_type() const override;
 
-	virtual void set_service_type(const String &type) = 0;
-	virtual String get_service_type() const = 0;
+	virtual void set_igd_control_url(const String &url) override;
+	virtual String get_igd_control_url() const override;
 
-	virtual void set_igd_control_url(const String &url) = 0;
-	virtual String get_igd_control_url() const = 0;
+	virtual void set_igd_service_type(const String &type) override;
+	virtual String get_igd_service_type() const override;
 
-	virtual void set_igd_service_type(const String &type) = 0;
-	virtual String get_igd_service_type() const = 0;
+	virtual void set_igd_our_addr(const String &addr) override;
+	virtual String get_igd_our_addr() const override;
 
-	virtual void set_igd_our_addr(const String &addr) = 0;
-	virtual String get_igd_our_addr() const = 0;
+	virtual void set_igd_status(IGDStatus status) override;
+	virtual IGDStatus get_igd_status() const override;
 
-	virtual void set_igd_status(IGDStatus status) = 0;
-	virtual IGDStatus get_igd_status() const = 0;
+	virtual bool is_valid_gateway() const override;
+	virtual String query_external_address() const override;
+	virtual int add_port_mapping(int port, int port_internal = 0, String desc = "", String proto = "UDP", int duration = 0) const override;
+	virtual int delete_port_mapping(int port, String proto = "UDP") const override;
 
-	virtual bool is_valid_gateway() const = 0;
-	virtual String query_external_address() const = 0;
-	virtual int add_port_mapping(int port, int port_internal = 0, String desc = "", String proto = "UDP", int duration = 0) const = 0;
-	virtual int delete_port_mapping(int port, String proto = "UDP") const = 0;
-
-	UPNPDevice() {}
-	virtual ~UPNPDevice() {}
+	UPNPDeviceMiniUPNP() {}
+	virtual ~UPNPDeviceMiniUPNP() {}
 };
 
-VARIANT_ENUM_CAST(UPNPDevice::IGDStatus)
+#endif // WEB_ENABLED
 
-#endif // UPNP_DEVICE_H
+#endif // UPNP_DEVICE_MINIUPNP_H
