+++ modules/upnp/upnp_miniupnp.h
@@ -1,5 +1,5 @@
 /**************************************************************************/
-/*  upnp_device.h                                                         */
+/*  upnp_miniupnp.h                                                       */
 /**************************************************************************/
 /*                         This file is part of:                          */
 /*                             GODOT ENGINE                               */
@@ -28,67 +28,66 @@
 /* SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                 */
 /**************************************************************************/
 
-#ifndef UPNP_DEVICE_H
-#define UPNP_DEVICE_H
+#ifndef UPNP_MINIUPNP_H
+#define UPNP_MINIUPNP_H
 
-#include "core/object/ref_counted.h"
+#ifndef WEB_ENABLED
 
-class UPNPDevice : public RefCounted {
-	GDCLASS(UPNPDevice, RefCounted);
+#include "upnp.h"
 
-protected:
-	static void _bind_methods();
+#include <miniupnpc.h>
 
-	static UPNPDevice *(*_create)(bool p_notify_postinitialize);
+class UPNPMiniUPNP : public UPNP {
+	GDCLASS(UPNPMiniUPNP, UPNP);
+
+private:
+	static UPNP *_create(bool p_notify_postinitialize) { return static_cast<UPNP *>(ClassDB::creator<UPNPMiniUPNP>(p_notify_postinitialize)); }
+
+	String discover_multicast_if = "";
+	int discover_local_port = 0;
+	bool discover_ipv6 = false;
+
+	Vector<Ref<UPNPDevice>> devices;
+
+	bool is_common_device(const String &dev) const;
+	void add_device_to_list(UPNPDev *dev, UPNPDev *devlist);
+	void parse_igd(Ref<UPNPDevice> dev, UPNPDev *devlist);
+	char *load_description(const String &url, int *size, int *status_code) const;
 
 public:
-	enum IGDStatus {
-		IGD_STATUS_OK,
-		IGD_STATUS_HTTP_ERROR,
-		IGD_STATUS_HTTP_EMPTY,
-		IGD_STATUS_NO_URLS,
-		IGD_STATUS_NO_IGD,
-		IGD_STATUS_DISCONNECTED,
-		IGD_STATUS_UNKNOWN_DEVICE,
-		IGD_STATUS_INVALID_CONTROL,
-		IGD_STATUS_MALLOC_ERROR,
-		IGD_STATUS_UNKNOWN_ERROR,
-	};
+	static void make_default();
 
-	static UPNPDevice *create(bool p_notify_postinitialize = true) {
-		if (!_create) {
-			return nullptr;
-		}
-		return _create(p_notify_postinitialize);
-	}
+	static int upnp_result(int in);
 
-	virtual void set_description_url(const String &url) = 0;
-	virtual String get_description_url() const = 0;
+	virtual int get_device_count() const override;
+	virtual Ref<UPNPDevice> get_device(int index) const override;
+	virtual void add_device(Ref<UPNPDevice> device) override;
+	virtual void set_device(int index, Ref<UPNPDevice> device) override;
+	virtual void remove_device(int index) override;
+	virtual void clear_devices() override;
 
-	virtual void set_service_type(const String &type) = 0;
-	virtual String get_service_type() const = 0;
+	virtual Ref<UPNPDevice> get_gateway() const override;
 
-	virtual void set_igd_control_url(const String &url) = 0;
-	virtual String get_igd_control_url() const = 0;
+	virtual int discover(int timeout = 2000, int ttl = 2, const String &device_filter = "InternetGatewayDevice") override;
 
-	virtual void set_igd_service_type(const String &type) = 0;
-	virtual String get_igd_service_type() const = 0;
+	virtual String query_external_address() const override;
 
-	virtual void set_igd_our_addr(const String &addr) = 0;
-	virtual String get_igd_our_addr() const = 0;
+	virtual int add_port_mapping(int port, int port_internal = 0, String desc = "", String proto = "UDP", int duration = 0) const override;
+	virtual int delete_port_mapping(int port, String proto = "UDP") const override;
 
-	virtual void set_igd_status(IGDStatus status) = 0;
-	virtual IGDStatus get_igd_status() const = 0;
+	virtual void set_discover_multicast_if(const String &m_if) override;
+	virtual String get_discover_multicast_if() const override;
 
-	virtual bool is_valid_gateway() const = 0;
-	virtual String query_external_address() const = 0;
-	virtual int add_port_mapping(int port, int port_internal = 0, String desc = "", String proto = "UDP", int duration = 0) const = 0;
-	virtual int delete_port_mapping(int port, String proto = "UDP") const = 0;
+	virtual void set_discover_local_port(int port) override;
+	virtual int get_discover_local_port() const override;
 
-	UPNPDevice() {}
-	virtual ~UPNPDevice() {}
+	virtual void set_discover_ipv6(bool ipv6) override;
+	virtual bool is_discover_ipv6() const override;
+
+	UPNPMiniUPNP() {}
+	virtual ~UPNPMiniUPNP() {}
 };
 
-VARIANT_ENUM_CAST(UPNPDevice::IGDStatus)
+#endif // WEB_ENABLED
 
-#endif // UPNP_DEVICE_H
+#endif // UPNP_MINIUPNP_H
